Automatic anchor generation for headers.
Headers automatically generate anchors just like they do for automatic
table-of-contents generation. In addition, if an implicit link is not
explicitly defined anywhere, then an automatically generated header
anchor is assumed.

The only method of enabling/disabling this option is via a static global
variable, `AutoAnchorHeaders`, defined in generate.c.