Support Markdown-formatted text inside HTML blocks
Previously <div> elements couldn’t wrap Markdown-formatted content
inside. This is because <div> is a block element and plain Markdown does
not format the content of such.

With this change the behavior for HTML block elements is configurable
now. To enable Markdown processing add a "markdown" attribute with a
value of "1" or "span" to the element. An empty SGML attribute, as in
"<div markdown>" also enables Markdown processing. To disable pass a
value of "0" or "block".

This feature was introduced by PHP Markdown Extra.

Limitations:

* In contrast to PHP Markdown Extra this doesn't permit to disable
  Markdown processing for inline elements yet.
* The value of this attribute is not inherited by child elements yet.