Provide a stable output for email address across runs
Email addresses were previously mangled randomly. This prevented two runs of
discount from providing the same output. This is problematic for software
using discount and comparing its output to some reference, or for software which
wants to generate documentation in a reproducible manner.

We can assume that the current code is trying to defeat an email address
scraper that supports one encoding but not the other. Simply alternating
between the two encodings would accomplish the same effect.

We thus change the behavior of the `mangle()` function to alternate the
encoding based on the position in the string instead of randomly. As there is
no other users of the `COINTOSS()` macro, we can remove its definition from the
configure script.