Fix 0xff-bytes being interpreted as EOF
On architectures with signed bytes, a 0xff-byte has value -1, which is also the
value of EOF. Fix this by casting to unsigned char in the right places.

I have no idea if the changes to the test are correct, but previously the
0xff-byte was interpreted as EOF and parsing stopped. After the code change, it
is no longer mis-interpreted as EOF and thus maybe_tag_or_link() finds the '>'
that is part of '</>' after the 0xff-byte and decides that things do not have to
be escaped.

Fixes: https://github.com/Orc/discount/issues/151
Signed-off-by: Uli Schlachter <psychon@znc.in>