Libtoolize for portability across pkgsrc platforms.
This fixes the self-tests with --shared on OS X, and generally makes
discount more portable. We've had this patch in pkgsrc since March
and I've used discount since then on at least NetBSD, OS X, Linux,
and Tribblix.

A definition of @INSTALL_LIB@ needs to be provided for this to work.