notspecial: do a positive check. (fixes plan9, improves others)
Currently, notspecial() checks for a list of special types of
file, instead of checking if the file is regular. It seems like
checking for what we want is likely to be more useful than
checking for things we want to avoid -- especially since the
list of special things is far more likely to grow.