tests/tabstop.t: Literal newline for tr(1)
Plan 9 tr(1) represents a newline with a literal newline. The \n
sequence yields the letter "n". So in tests/tabstop.t:7 the eval
yields "tests/tabstop.t: markdown:ndiscountn3nTAB=7nGHC=INPUT: not found"
from a make tests run.

If you do just the eval in straight ape/psh, no interceding mk:
    % ape/psh
    $ ./markdown -V
    markdown: discount 3 TAB=7 GHC=INPUT
    $ eval `./markdown -V | tr ' ' '\n' | grep TAB`
    /bin/sh: markdown:ndiscountn3nTAB=7nGHC=INPUT: not found

This change makes the eval and thus the tabstop test work on
Plan 9, but I'm not sure beyond a quick test if it is OK to
throw a literal newline at GNU tr.