Fix isspace() usage.
Both ANSI and POSIX define that the arguments for isspace()
need to be EOF or valid as 'unsigned char'.

On many platforms, 'char' is 'signed char'. When parsing e.g. UTF-8
documents on such a platform, invalid values like -61 are passed to
isspace.

This causes core dumps when building the documentation for
libretro-mgba on NetBSD.